Add translate_index
Globus Search prefers (and will eventually require) users to use index UUIDs instead of names. As names are generally easier to remember, having a translator between the two is handy. There is a dictionary of common/known name=>UUID mappings in Toolbox, but unknown names will be looked up using the SearchClient.