Unify logins
Greatly expand the scope of the toolbox login functions (login, confidential_login, anonymous_login) to allow the user to specify any Globus scope and get back an authorizer (where possible), or specify a known service and get back an appropriate Client or authorizer. The internals of anonymous_login have changed, but the behavior is not significantly different; there is no way to authenticate to an arbitrary scope anonymously.