Set standard timeout on all created clients
Previously, MDF services had an issue with custom_transfer timing out on the directory check. This was solved by temporarily raising the timeout on the TransferClient. However, now it is apparently necessary to also raise the timeout on SearchClients. Since this seems to be a pattern, I'm setting a standard timeout of 5 minutes and applying it to all globus_sdk clients created through Toolbox. This change should not have a meaningful impact on general use, but will make client use easier in certain edge cases.

Additionally, since the custom_transfer timeout-raise code is now redundant, I've removed it.