Format UUIDs into URNs in ACLs for valid GMeta, in format_gmeta()
The logic here just double-formats bare UUID into the two possible categories of URN: User and Group. This would be unnecessary if all Globus UUIDs were URNs (as there would not be some bare UUIDs and some URN-UUIDs), but currently that is not the case. Doubling up on the UUIDs in this way should not have any visible impact, as UUIDs only point to one object (and are not reused between User IDs and Group IDs).