Add FuncX scope to known auth scopes
For the record: The list of known scopes does NOT limit the auth functionality of Toolbox. Any valid Globus Auth scope can be specified and will be used. The known scope list is only a convenience feature to give memorable names to certain scopes that are commonly used by Toolbox's users.